+++ Java-veebirakendus/src/main/java/ee/itcollege/group04_2011/entities/AmetPiiripunkti.java
@@ -34,7 +34,7 @@ public class AmetPiiripunkti implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="AMET_PIIRIPUNKTIS_ID")
-	private int ametPiiripunktisId;
+	private Long ametPiiripunktisId;
 
     @Temporal( TemporalType.DATE)
 	private Date alates;
@@ -76,11 +76,11 @@ public class AmetPiiripunkti implements Serializable {
     public AmetPiiripunkti() {
     }
 
-	public int getAmetPiiripunktisId() {
+	public Long getAmetPiiripunktisId() {
 		return this.ametPiiripunktisId;
 	}
 
-	public void setAmetPiiripunktisId(int ametPiiripunktisId) {
+	public void setAmetPiiripunktisId(Long ametPiiripunktisId) {
 		this.ametPiiripunktisId = ametPiiripunktisId;
 	}
