+++ Java-veebirakendus/src/main/java/ee/itcollege/group04_2011/entities/Auaste.java
@@ -30,7 +30,7 @@ public class Auaste implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="AUASTE_ID")
-	private int auasteId;
+	private Long auasteId;
 
 	private String avaja;
 
@@ -60,11 +60,11 @@ public class Auaste implements Serializable {
     public Auaste() {
     }
 
-	public int getAuasteId() {
+	public Long getAuasteId() {
 		return this.auasteId;
 	}
 
-	public void setAuasteId(int auasteId) {
+	public void setAuasteId(Long auasteId) {
 		this.auasteId = auasteId;
 	}
