+++ Java-veebirakendus/src/main/java/ee/itcollege/group04_2011/entities/OrgYksus.java
@@ -34,7 +34,7 @@ public class OrgYksus implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="ORG_YKSUS_ID")
-	private int orgYksusId;
+	private Long orgYksusId;
 
     @Temporal( TemporalType.DATE)
 	private Date alates;
@@ -80,11 +80,11 @@ public class OrgYksus implements Serializable {
     public OrgYksus() {
     }
 
-	public int getOrgYksusId() {
+	public Long getOrgYksusId() {
 		return this.orgYksusId;
 	}
 
-	public void setOrgYksusId(int orgYksusId) {
+	public void setOrgYksusId(Long orgYksusId) {
 		this.orgYksusId = orgYksusId;
 	}
